Move temporary file handler to categories
